Task 16 Code refactoring
