--- src/main/resources/am-service.properties
@@ -1,2 +1,0 @@
-site.base.url=https://www.amwine.ru/catalog/vino
-site.user.agent=Chrome/4.0.249.0 Safari/532.5
